added substring to UString

+++ src/aurora/talktable.cpp
@@ -23,7 +23,7 @@
  */
 
 #include "src/common/util.h"
-#include "src/common/stream.h"
+#include "src/common/readstream.h"
 
 #include "src/aurora/aurorafile.h"
 #include "src/aurora/talktable.h"
@@ -45,7 +45,7 @@ TalkTable *TalkTable::load(Common::SeekableReadStream *tlk, Common::Encoding enc
 	if (!tlk)
 		return 0;
 
-	uint32 pos = tlk->pos();
+	size_t pos = tlk->pos();
 
 	uint32 id, version;
 	bool utf16le;
