+++ src/engines/nwn/gui/widgets/buttonsgroup.cpp
@@ -64,7 +64,7 @@ void ButtonsGroup::setActive(WidgetButton *button) {
 	     it != _buttonsList.end(); ++it) {
 		// First we check if the button belongs to this group.
 		if (*it == button) {
-			_helpBox->setText("fnt_galahad14", _helpTexts[std::distance(_buttonsList.begin(), it)], 1.0);
+			_helpBox->setText("fnt_galahad14", _helpTexts[std::distance(_buttonsList.begin(), it)], 1.0f);
 			_helpBox->setTitle("fnt_galahad14", _helpTitles[std::distance(_buttonsList.begin(), it)]);
 
 			if ((*it)->getMode() == WidgetButton::kModeUnchanged)
@@ -86,7 +86,7 @@ void ButtonsGroup::setActive(WidgetButton *button) {
 	}
 }
 
-void ButtonsGroup::setActive(uint button) {
+void ButtonsGroup::setActive(size_t button) {
 	if (button >= _buttonsList.size())
 		error("Wrong ID buttons");
 
@@ -99,7 +99,7 @@ void ButtonsGroup::setAllInactive() {
 		(*it)->setPressed(false);
 }
 
-uint ButtonsGroup::getChoice() const {
+size_t ButtonsGroup::getChoice() const {
 	for (std::vector<WidgetButton *>::const_iterator it = _buttonsList.begin(); it != _buttonsList.end(); ++it) {
 		if ((*it)->isPressed()) {
 			return std::distance(_buttonsList.begin(), it);
