+++ src/sound/decoders/aac.cpp
@@ -24,8 +24,10 @@
  *  Decoding AAC.
  */
 
+#include <cstring>
+
 #include "src/common/error.h"
-#include "src/common/stream.h"
+#include "src/common/memreadstream.h"
 
 #include "src/sound/audiostream.h"
 #include "src/sound/decoders/aac.h"
@@ -86,8 +88,8 @@ AACDecoder::~AACDecoder() {
 
 AudioStream *AACDecoder::decodeFrame(Common::SeekableReadStream &stream) {
 	// read everything into a buffer
-	uint32 inBufferPos = 0;
-	uint32 inBufferSize = stream.size();
+	size_t inBufferPos = 0;
+	size_t inBufferSize = stream.size();
 	byte *inBuffer = new byte[inBufferSize];
 	stream.read(inBuffer, inBufferSize);
